requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-classic
python-dotenv
fastapi
uvicorn
streamlit
duckduckgo-search
faiss-cpu
ddg
